sports-ball-classification/inference/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
Pillow==10.2.0
tensorflow>=2.15.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
numpy>=1.24.0
psycopg2-binary==2.9.9
